+++ src/main/java/uk/co/techblue/alfresco/dto/user/User.java
@@ -26,7 +26,7 @@ import uk.co.techblue.alfresco.jackson.ISO8601DateDeserializer;
 /**
  * The Class User.
  */
-public class User extends BaseDto {
+public class User extends BaseUser {
 
     /** The Constant serialVersionUID. */
     private static final long serialVersionUID = -3009100093342107112L;
@@ -46,18 +46,6 @@ public class User extends BaseDto {
     /** The avatar. */
     @JsonProperty("avatar")
     private String avatar;
-
-    /** The first name. */
-    @JsonProperty("firstName")
-    private String firstName;
-
-    /** The last name. */
-    @JsonProperty("lastName")
-    private String lastName;
-
-    /** The jobtitle. */
-    @JsonProperty("jobtitle")
-    private String jobtitle;
 
     /** The organization. */
     @JsonProperty("organization")
@@ -200,60 +188,6 @@ public class User extends BaseDto {
      */
     public void setEnabled(final boolean enabled) {
         this.enabled = enabled;
-    }
-
-    /**
-     * Gets the first name.
-     * 
-     * @return the first name
-     */
-    public String getFirstName() {
-        return firstName;
-    }
-
-    /**
-     * Sets the first name.
-     * 
-     * @param firstName the new first name
-     */
-    public void setFirstName(final String firstName) {
-        this.firstName = firstName;
-    }
-
-    /**
-     * Gets the last name.
-     * 
-     * @return the last name
-     */
-    public String getLastName() {
-        return lastName;
-    }
-
-    /**
-     * Sets the last name.
-     * 
-     * @param lastName the new last name
-     */
-    public void setLastName(final String lastName) {
-        this.lastName = lastName;
-    }
-
-    /**
-     * Gets the jobtitle.
-     * 
-     * @return the jobtitle
-     */
-    public String getJobtitle() {
-        return jobtitle;
-    }
-
-    /**
-     * Sets the jobtitle.
-     * 
-     * @param jobtitle the new jobtitle
-     */
-    public void setJobtitle(final String jobtitle) {
-        this.jobtitle = jobtitle;
     }
 
     /**
